Rename files and add a debug log for case ref
